requirements/base.txt
Django==3.0.7
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.2.0
psycopg2==2.8.5
